+++ src/Dialog/index.ts
@@ -1,2 +1,3 @@
 export * from './Dialog';
 export * from './useDialog';
+export * from './DialogHeader';