+++ src/Drawer/index.ts
@@ -1,2 +1,3 @@
 export * from './Drawer';
 export * from './useDrawer';
+export * from './DrawerHeader';